requirements.txt
openai>=0.27.0
pinecone-client>=2.2.0
python-dotenv>=0.21.0
tqdm>=4.64.0